[DIRTY DRAFT] rasterize polygons

+++ src/main/java/engine/ColorRGB.java
@@ -1,9 +1,9 @@
 package engine;
 
 public class ColorRGB {
-    public double red, green, blue;
+    public byte red, green, blue;
 
-    public ColorRGB(double red, double green, double blue) {
+    public ColorRGB(byte red, byte green, byte blue) {
         this.red = red;
         this.green = green;
         this.blue = blue;
